requirements.txt
# Agente DGP - Dirección General de Profesiones (SEP)
# LangChain y LangSmith
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
langsmith>=0.1.0
langchain-groq

# RAG y Embeddings
langchain-community>=0.3.0
chromadb>=0.5.0
sentence-transformers>=2.2.0

# API y servidor web
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Utilidades
python-dotenv>=1.0.0
